+++ UpgradeEmptyCabins/Framework/ModConfig.cs
@@ -1,6 +1,0 @@
-namespace UpgradeEmptyCabins.Framework;
-
-internal class ModConfig
-{
-    public bool InstantBuild { get; set; } = false;
-}